chore: refactor pipeline methods
