+++ dozer-ingestion/src/connectors/files/mod.rs
@@ -1,0 +1,4 @@
+pub mod connector;
+
+#[cfg(test)]
+mod tests;
