+++ dozer-orchestrator/src/lib.rs
@@ -47,7 +47,6 @@ pub trait Orchestrator {
 }
 
 // Re-exports
-use dozer_ingestion::connectors::{TableInfo, ValidationResults};
 pub use dozer_ingestion::{connectors::get_connector, errors::ConnectorError};
 
 pub use dozer_types::models::connection::Connection;
@@ -63,18 +62,6 @@ async fn flatten_joinhandle(
         Err(err) => Err(OrchestrationError::InternalError(Box::new(err))),
     }
 }
-
-pub fn validate(input: Connection, tables: Option<Vec<TableInfo>>) -> Result<(), ConnectorError> {
-    get_connector(input)?.validate(tables)
-}
-
-pub fn validate_schema(
-    input: Connection,
-    tables: &[TableInfo],
-) -> Result<ValidationResults, ConnectorError> {
-    get_connector(input)?.validate_schemas(tables)
-}
-
 pub fn set_panic_hook() {
     panic::set_hook(Box::new(move |panic_info| {
         // All the orchestrator errors are captured here
