+++ dozer-orchestrator/src/pipeline/mod.rs
@@ -1,6 +1,10 @@
+mod builder;
 pub mod connector_source;
 mod sinks;
 pub mod source_builder;
 mod streaming_sink;
+pub mod validate;
+pub use builder::PipelineBuilder;
 pub use sinks::{CacheSink, CacheSinkFactory, CacheSinkSettings};
 pub(crate) use streaming_sink::StreamingSinkFactory;
+mod basic_processor_factory;
