+++ dozer-orchestrator/src/simple/mod.rs
@@ -1,8 +1,5 @@
 mod executor;
 pub mod orchestrator;
 pub use orchestrator::SimpleOrchestrator;
-mod basic_processor_factory;
-mod direct_cache_pipeline;
-
 #[cfg(test)]
 mod tests;
